add paper scenario, various bug fixes

+++ src/mmwave/model/mmwave-enb-mac.h
@@ -99,7 +99,7 @@ public:
     uint8_t m_ccId; //!< The Component Carrier ID
     MmWaveMacSchedSapUser::SchedConfigIndParameters
         m_indParam; //!< Structure containing all the scheduling decisions
-    uint32_t m_cellId;
+    uint16_t m_cellId;
   };
 
   /*
